Refactor: Restructure code related to saving text in the local storage
Added a new paramater to the custom hook useLocalStoredText that will contain the type of text stored. This is done to store different types of text while using the same unique identifier to be part of the local key. This functionality will be used for the answer and prompt text. Also encapsulated code related to the title under TitleInput.