Fix and Feature: Fix title section bug and add problem input feature
Previously, the title section did not properly display the title as, when the program first ran, the title did not display in the section dashboard until the section was entered. Updated Section.jsx to retrieve the title stored in the local storage and to display the title to resolve the issue. Also encountered another bug with the where the section title would display the removed section, after the section was removed. Updated SectionContent.jsx to keep the TitleInput in the program, without showing it, so that the component does not render again, which lead to issues with its useEffect that lead to this problem. Also added feature to save the text inputted in the answer and prompt textareas through the local storage.