Fix: Fix error when entering a section
Previously, an error would appear when entering a section after removing a section. This occured because entering a section would run code that would set the background of the previous selected section in the sections dashboard to black. Since the section is deleted however, the stored reference in selectedSectionRef would have a reference that would not exist. Fixed this issue by setting selectedSectionRef to null when a section is removed.