Fix and Feature: Fix bug where multiple choice options and the prompt did not match and added feature to multiple choice
Previously, among the multiple choice options, the answer to the prompt was not available. Resolved bug by ensuring that the handleShuffle function ran specifically when the value of the answer variable updated. To do this, the useEffect that sets the value of the stop variable used the answer variable as a dependency. Also added feature to allow the user to check their multiple choice answer.