Fix: Fix return to dashboard bug
Previously, after refreshing the test page then clicking on the return to the dashboard button, the section previously selected would no longer be selected. Resolved bug by saving important properties of a section in the local storage, so that the needed information for the general dashboard when exiting the test page would be saved regardless if the test page was refreshed.