Fix: Fix return dashboard bug and multiple choice bug
Previously, an error would be thrown if the user clicked on the return button on the CompleteTest page. Resolved bug by passing the missing arguments that were necessary for the ReturnToDashboardButton component to work. Also, previously, if all the answer inputs and the prompt inputs were the same for a section and on the test page, the user encounters multiple problems that are multiple choice, the previously selected choice would appear to stay selected on the next multiple choice problem. Resolved this bug by updating Choice.jsx to display the previously selected choice as unselected if the next problem is a multiple choice problem.