Fix: Fix bug regarding test statistics summary
Previously, the test statistics improperly displayed the percentage the user answered correctly as it sometimes did not round to the nearest hundredth. Updated CompleteTest to calculate and multiply the decimal by 100 within the parenthesis that is used for the toFixed function to ensure all the numbers needed for the calculation are the same type. Also changed the rounding to round to the nearest tenth as the percentage appears less overwhelming by rounding closer to the decimal point.